requirements.txt
Pillow>=10.0.0
pytubefix>=6.0.0
faster-whisper>=1.0.0
